CCD-2493: ccd-definition-store-api Validate the scope of transactions in CCD     [CCD-2493] (https://tools.hmcts.net/jira/browse/CCD-2493).
