+++ elastic-search-support/src/main/java/uk/gov/hmcts/ccd/definition/store/elastic/endpoint/ElasticsearchIndexController.java
@@ -11,7 +11,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.http.HttpStatus;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -48,7 +47,6 @@ public class ElasticsearchIndexController {
         this.elasticGlobalSearchListener = elasticGlobalSearchListener;
     }
 
-    @Transactional
     @PostMapping(ELASTIC_INDEX_URI)
     @ResponseStatus(HttpStatus.CREATED)
     @ApiOperation("Builds the Elasticsearch indices for all known case types, using their latest version. This API can "
@@ -72,7 +70,6 @@ public class ElasticsearchIndexController {
         return new IndicesCreationResult(caseTypesToIndex);
     }
 
-    @Transactional
     @PostMapping(GS_ELASTIC_INDEX_URI)
     @ResponseStatus(HttpStatus.CREATED)
     @ApiOperation("Builds the Elasticsearch index for Global Search.")
